refactor(utils): keep the validator in utils.validation, defer the containers import
Follows @HugoDelatte's second review on #279.

`validate_asset_panel` stays where users already import it from. The
`skfolio.base` relocation, its `__all__` entry, the `utils/validation.py`
deprecation shim and all 46 rewritten caller imports are reverted, so the
public import path is unchanged and there is nothing to deprecate:

    from skfolio.utils.validation import validate_asset_panel

The `utils -> containers` edge is broken instead by importing `AssetPanel` and
`AssetPanelView` inside the function, with a comment saying why. Annotations
are already postponed, so the signature is covered by a `TYPE_CHECKING` import
and the runtime import happens only when the validator is called.

`validate_asset_panel` and `validate_cross_sectional_data` are documented under
a new `skfolio.utils.validation` section in `docs/api.rst`; the module had no
API page before. Validator tests return to `tests/test_utils/test_validation.py`
unchanged, and `tests/test_base.py` is dropped.

The neutralization moves are unchanged from the previous round.

`skfolio.utils` now imports only `_constants`, `exceptions` and `typing`, and
the package graph has no cycles — down from six on `main`
(`utils <-> containers`, `utils <-> linear_model`, `utils <-> preprocessing`).
Full suite: 6967 passed. `ruff check` and `ruff format --check` clean.
