+++ docs/api.rst
@@ -221,16 +221,6 @@ Classes
     base.BaseAssetPanelTransformer
     base.BaseComposition
 
-Functions
----------
-.. currentmodule:: skfolio
-
-.. autosummary::
-    :toctree: generated/
-    :template: function.rst
-
-    base.validate_asset_panel
-
 
 .. _optimization_base_ref:
 
@@ -1126,6 +1116,22 @@ Functions
     stats.squared_standardized_euclidean_dist
     stats.symmetric_step_up_matrix
     stats.symmetrize
+
+.. _validation_ref:
+
+:mod:`skfolio.utils.validation`: Validation
+===========================================
+
+Functions
+---------
+.. currentmodule:: skfolio.utils
+
+.. autosummary::
+    :toctree: generated/
+    :template: function.rst
+
+    validation.validate_asset_panel
+    validation.validate_cross_sectional_data
 
 .. _distribution_ref:
