+++ src/skfolio/alpha/_fixed_weighted_alpha.py
@@ -12,7 +12,6 @@ import skfolio.typing as skt
 from skfolio._constants import _EXPOSURES, _IDIO_VARIANCES, _PASSTHROUGH
 from skfolio.alpha import ForecastUnit
 from skfolio.alpha._base import BaseAlpha, BaseAlphaDescriptorComposition
-from skfolio.base import validate_asset_panel
 from skfolio.containers import AssetPanel
 from skfolio.descriptor import BaseDescriptor
 from skfolio.typing import FloatArray
@@ -21,6 +20,7 @@ from skfolio.utils.tools import (
     _validate_positive_real,
     _validate_unit_interval,
 )
+from skfolio.utils.validation import validate_asset_panel
 
 _FITTED_ATTR = "alpha_"
