+++ src/skfolio/descriptor/_short_interest/_days_to_cover.py
@@ -8,7 +8,6 @@ from __future__ import annotations
 
 import numpy as np
 
-from skfolio.base import validate_asset_panel
 from skfolio.containers import AssetPanel
 from skfolio.descriptor._base import BaseDescriptor
 from skfolio.typing import FloatArray
@@ -18,6 +17,7 @@ from skfolio.utils.tools import (
     _validate_positive_real,
     half_life_to_decay_factor,
 )
+from skfolio.utils.validation import validate_asset_panel
 
 _FITTED_ATTR = "days_to_cover_"
