+++ src/skfolio/utils/validation.py
@@ -6,18 +6,18 @@
 
 from __future__ import annotations
 
-import warnings
-from typing import Literal
+from typing import TYPE_CHECKING, Literal
 
 import numpy as np
 import sklearn.utils.validation as skv
 from sklearn.utils._tags import get_tags
 
-from skfolio.typing import ArrayLike, FloatArray
+from skfolio.typing import AnyArray, ArrayLike, FloatArray
 
-# `validate_asset_panel` is deprecated and resolved by the module-level
-# `__getattr__` below, so wildcard imports keep working until v2.0.
-__all__ = ["validate_asset_panel", "validate_cross_sectional_data"]  # noqa: F822
+if TYPE_CHECKING:
+    from skfolio.containers import AssetPanel, AssetPanelView
+
+__all__ = ["validate_asset_panel", "validate_cross_sectional_data"]
 
 
 def validate_cross_sectional_data(
@@ -194,26 +194,267 @@ def validate_cross_sectional_data(
     return X_validated, y_validated, w_validated
 
 
-# TODO remove deprecated validate_asset_panel re-export in v2.0
-def __getattr__(name: str):
-    """Resolve the deprecated `validate_asset_panel` re-export.
+def validate_asset_panel(
+    _estimator,
+    /,
+    asset_panel: AssetPanel | AssetPanelView,
+    required_fields: list[str] | None = None,
+    reserved_fields: list[str] | None = None,
+    finite_or_nan: list[str] | None = None,
+    finite_when_active: list[str] | None = None,
+    strictly_positive_or_nan: list[str] | None = None,
+    strictly_positive_when_active: list[str] | None = None,
+    non_negative_or_nan: list[str] | None = None,
+    reset: bool = True,
+    copy: bool = False,
+) -> AssetPanel | AssetPanelView:
+    """Validate an AssetPanel and set estimator metadata attributes.
 
-    `validate_asset_panel` moved to :mod:`skfolio.base`, which cannot be imported at
-    module level here because :mod:`skfolio.base` depends on :mod:`skfolio.utils`. The
-    import is therefore deferred until the deprecated name is actually accessed.
+    This function validates that the panel contains required fields, doesn't contain
+    reserved ones and sets standard metadata attributes on the estimator.
 
-    .. deprecated::
-        `skfolio.utils.validation.validate_asset_panel` is deprecated and will be
-        removed in version 2.0. Import it from :mod:`skfolio.base` instead.
+    Parameters
+    ----------
+    _estimator : estimator instance
+        The estimator on which to set validation attributes.
+
+    asset_panel : AssetPanel or AssetPanelView
+        AssetPanel. AssetPanel already validates that all fields have consistent shapes
+        and consistent masks, allowing this validation to be lightweight.
+
+    required_fields : list of str, optional
+        Fields that must be present in the AssetPanel. If any are missing, a ValueError
+        is raised.
+
+    reserved_fields : list of str, optional
+        Fields that must NOT be present in the AssetPanel. These are typically names the
+        estimator will create internally. If any are found, a ValueError is raised.
+
+    finite_or_nan : list of str, optional
+        Fields whose values must be finite or NaN.
+
+    finite_when_active : list of str, optional
+        Fields whose values must be finite wherever `active_mask` is True. This is
+        typically used for fields like "market_cap" that must be forward-filled for
+        holidays before constructing the AssetPanel.
+
+    strictly_positive_or_nan : list of str, optional
+        Fields whose values must be strictly positive and finite, or NaN.
+
+    strictly_positive_when_active : list of str, optional
+        Fields whose values must be strictly positive and finite wherever
+        `active_mask` is True.
+
+    non_negative_or_nan : list of str, optional
+        Fields whose values must be non-negative and finite, or NaN.
+
+    reset : bool, default=True
+        If True, sets metadata attributes on the estimator:
+        - `asset_names_`: array of asset identifiers
+        - `n_assets_`: number of assets
+
+        If False, validates that existing attributes match the panel.
+
+    copy : bool, default=False
+        If True, returns a shallow copy of the panel (new dict containers with shared
+        arrays). Use this when the estimator needs to add fields without mutating the
+        user's input.
+
+    Returns
+    -------
+    AssetPanel or AssetPanelView
+        The validated panel, or a shallow copy if `copy=True`.
+
+    Raises
+    ------
+    TypeError
+        If asset_panel is not an AssetPanel or AssetPanelView.
+
+    ValueError
+        If required fields are missing, reserved fields are present, a
+        field validation rule is violated or (when reset=False) panel doesn't match
+        stored metadata.
     """
-    if name == "validate_asset_panel":
-        from skfolio.base import validate_asset_panel
+    # Normalize empty lists to None.
+    if required_fields is not None and len(required_fields) == 0:
+        required_fields = None
+    if reserved_fields is not None and len(reserved_fields) == 0:
+        reserved_fields = None
+    if finite_or_nan is not None and len(finite_or_nan) == 0:
+        finite_or_nan = None
+    if finite_when_active is not None and len(finite_when_active) == 0:
+        finite_when_active = None
+    if strictly_positive_or_nan is not None and len(strictly_positive_or_nan) == 0:
+        strictly_positive_or_nan = None
+    if (
+        strictly_positive_when_active is not None
+        and len(strictly_positive_when_active) == 0
+    ):
+        strictly_positive_when_active = None
+    if non_negative_or_nan is not None and len(non_negative_or_nan) == 0:
+        non_negative_or_nan = None
 
-        warnings.warn(
-            "skfolio.utils.validation.validate_asset_panel is deprecated and will "
-            "be removed in version 2.0. Import it from skfolio.base instead.",
-            FutureWarning,
-            stacklevel=2,
+    # Imported here rather than at module scope: `skfolio.containers` imports
+    # `skfolio.utils.tools`, so a module-level import would close a package-level
+    # cycle between `skfolio.utils` and `skfolio.containers`. Annotations are
+    # postponed, so the `TYPE_CHECKING` import above covers the signature.
+    from skfolio.containers import AssetPanel, AssetPanelView
+
+    if not isinstance(asset_panel, (AssetPanel, AssetPanelView)):
+        raise TypeError(
+            f"Must be an AssetPanel or AssetPanelView, got {type(asset_panel).__name__}"
         )
-        return validate_asset_panel
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+    field_names = list(asset_panel.keys())
+    asset_names = asset_panel.asset_names
+
+    # Check consistency with previous fit (if reset=False).
+    if not reset:
+        if not np.array_equal(_estimator.asset_names_, asset_names):
+            raise ValueError(
+                f"asset_names don't match. Expected {list(_estimator.asset_names_)}, "
+                f"got {list(asset_names)}."
+            )
+
+    # Check reserved fields.
+    if reserved_fields is not None:
+        reserved = set(reserved_fields) & set(field_names)
+        if reserved:
+            raise ValueError(
+                f"Reserved fields must be removed or renamed: {sorted(reserved)}."
+            )
+
+    # Check required fields.
+    if required_fields is not None:
+        missing = set(required_fields) - set(field_names)
+        if missing:
+            raise ValueError(
+                f"Required fields are missing: {sorted(missing)}. "
+                f"Available: {sorted(field_names)}."
+            )
+
+    # Check finite-or-NaN constraints.
+    if finite_or_nan is not None:
+        _check_fields_exist(finite_or_nan, field_names, "finite_or_nan")
+        for field in finite_or_nan:
+            values = asset_panel[field]
+            if values.ndim == 2:
+                bad = np.isinf(values)
+            else:
+                bad = np.isinf(values).any(axis=2)
+            if bad.any():
+                bad_obs = _bad_observation(bad)
+                raise ValueError(
+                    f'Field "{field}" contains infinite values '
+                    f"(first at observation index {bad_obs}). "
+                    f'"{field}" must contain finite values or NaN.'
+                )
+
+    # Check finite-when-active constraints.
+    if finite_when_active is not None:
+        _check_fields_exist(finite_when_active, field_names, "finite_when_active")
+        for field in finite_when_active:
+            values = asset_panel[field]
+            if values.ndim == 2:
+                bad = ~np.isfinite(values) & asset_panel.active_mask
+            else:
+                # 3-D field: check all components
+                bad = (~np.isfinite(values)).any(axis=2) & asset_panel.active_mask
+            if bad.any():
+                bad_obs = _bad_observation(bad)
+                raise ValueError(
+                    f'Field "{field}" contains NaN/inf for active assets '
+                    f"(first at observation index {bad_obs}). "
+                    f'"{field}" must be finite wherever `active_mask` is '
+                    f'True. Forward-fill "{field}" for holidays '
+                    f'or set `active_mask` to False until the first finite "{field}".'
+                )
+
+    # Check strictly-positive-or-NaN constraints.
+    if strictly_positive_or_nan is not None:
+        _check_fields_exist(
+            strictly_positive_or_nan, field_names, "strictly_positive_or_nan"
+        )
+        for field in strictly_positive_or_nan:
+            values = asset_panel[field]
+            if values.ndim == 2:
+                bad = ~np.isnan(values) & (~np.isfinite(values) | (values <= 0))
+            else:
+                bad = (~np.isnan(values) & (~np.isfinite(values) | (values <= 0))).any(
+                    axis=2
+                )
+            if bad.any():
+                bad_obs = _bad_observation(bad)
+                raise ValueError(
+                    f'Field "{field}" contains non-positive or infinite values '
+                    f"(first at observation index {bad_obs}). "
+                    f'"{field}" must contain strictly positive finite values or NaN.'
+                )
+
+    # Check strictly-positive-when-active constraints.
+    if strictly_positive_when_active is not None:
+        _check_fields_exist(
+            strictly_positive_when_active,
+            field_names,
+            "strictly_positive_when_active",
+        )
+        for field in strictly_positive_when_active:
+            values = asset_panel[field]
+            if values.ndim == 2:
+                bad = (~np.isfinite(values) | (values <= 0)) & asset_panel.active_mask
+            else:
+                bad = (~np.isfinite(values) | (values <= 0)).any(
+                    axis=2
+                ) & asset_panel.active_mask
+            if bad.any():
+                bad_obs = _bad_observation(bad)
+                raise ValueError(
+                    f'Field "{field}" contains non-finite or non-positive values for active '
+                    f"assets "
+                    f"(first at observation index {bad_obs}). "
+                    f'"{field}" must be strictly positive and finite wherever '
+                    f"`active_mask` is True."
+                )
+
+    # Check non-negative-or-NaN constraints.
+    if non_negative_or_nan is not None:
+        _check_fields_exist(non_negative_or_nan, field_names, "non_negative_or_nan")
+        for field in non_negative_or_nan:
+            values = asset_panel[field]
+            if values.ndim == 2:
+                bad = ~np.isnan(values) & (~np.isfinite(values) | (values < 0))
+            else:
+                bad = (~np.isnan(values) & (~np.isfinite(values) | (values < 0))).any(
+                    axis=2
+                )
+            if bad.any():
+                bad_obs = _bad_observation(bad)
+                raise ValueError(
+                    f'Field "{field}" contains negative values or infinite values '
+                    f"(first at observation index {bad_obs}). "
+                    f'"{field}" must contain non-negative finite values or NaN.'
+                )
+
+    # Set estimator metadata attributes.
+    if reset:
+        _estimator.asset_names_ = np.asarray(asset_names)
+        _estimator.n_assets_ = len(asset_names)
+
+    return asset_panel.copy() if copy else asset_panel
+
+
+def _check_fields_exist(
+    fields: list[str], field_names: list[str], arg_name: str
+) -> None:
+    """Check that all validation rule fields exist in the panel."""
+    for field in fields:
+        if field not in field_names:
+            raise ValueError(
+                f'{arg_name} lists "{field}", but that field is not in the '
+                f"AssetPanel. Available: {sorted(field_names)}."
+            )
+
+
+def _bad_observation(bad: AnyArray) -> int:
+    """Return the first observation index containing a validation failure."""
+    return int(np.where(bad.any(axis=1))[0][0])
